+++ CHANGELOG.md
@@ -15,6 +15,8 @@ All notable changes to this project will be documented in this file.
 
 - Moved 'forward fill' process from 'clean' job to 'impute' job. Moved tests for the same.
 
+- Set up estimates step for job role breakdown.
+
 ### Fixed